Add class Result and set up relationship with Analyses, Samples tables.
Proof of concept is now in place for the three tables in the
database. The insert_db script currently adds a new Analysis,
Sample and Result to each of the tables, and prints the effect
observed when a Sample record is added to an Analysis record.
Note that these records are all hardcoded as a POC and a CLI
still needs to be implemented.

Tests also still need to be written.