Fix create_db to dynamically generate Result model using  metric names from user
The script create_db now accepts a csv file which stipulates the names and
data types of metrics which the user wants in their database. This was
accomplished by making use of python's type() function, which magically creates
a new class when supplied with class name, class to inherit from, and attributes
to add to the class. This allows a user to choose which metrics they want to
include in their database upfront, as this differs between projects.

NOTE: that although this script works to create the appropriate database
with user supplied metric field, the class created dynamically only lives in
memory. This means that the insert_db script will now be broken as the class
Result does not exist. Will therefore have to find a different solution in a
future commit.

Also added a click CLI to both create_db and insert_db scripts.

Tests must still be written for everything.