Swap database backend to PostgreSQL
This represents a major change to the project. Although SQLAlchemy makes
swapping out the database backend fairly painless, the models have been
changed significantly. This makes the code fairly specific to PostgreSQL
going forward. Most notably is the inclusion of JSON types in fields.

Note: considerable setup may be required for setting up the database server.
This is in the early stages and we are still deciding on hosting plans.

Minor modifications to requirements files and CLI of both scripts.